backend/requirements.txt
Flask
psycopg2-binary
flask_cors
Werkzeug==1.0.1
flask_bcrypt
